Change files for typescript files
For now, types are setup correctly for index.ts. There is still work needed in every other files, and even in index.ts, some types need some work.

For now I have no way of testing if it works, but since types are only in typescript, nothing should be transpiled. I've never done this before so I am not 100% certain.